refactor code to new write.rs mod
